add the group memmber request

+++ Model/Basic/LLQQCategory.h
@@ -11,14 +11,14 @@
 
 @interface LLQQCategory : LLModelObject
 {
-    int index;       /*分组的唯一标识*/
-    int sort;        /* ? */
+    long index;       /*分组的唯一标识*/
+    long sort;        /* ? */
     NSString *name;  /*分组名称*/
     NSMutableDictionary *usersMap;  /*该组下所有的用户, key是QQ号的字符串形式 */
 }
 
-@property (nonatomic, assign) int index;  
-@property (nonatomic, assign) int sort;
+@property (nonatomic, assign) long index;  
+@property (nonatomic, assign) long sort;
 @property (nonatomic, copy) NSString *name;
 @property (nonatomic, retain) NSMutableDictionary *usersMap;
